+++ libraries/lychee/source/core/Specification.js
@@ -1,0 +1,39 @@
+
+lychee.Specification = typeof lychee.Specification !== 'undefined' ? lychee.Specification : (function(global) {
+
+
+
+	/*
+	 * IMPLEMENTATION
+	 */
+
+	const Composite = function(data) {
+
+		let settings = Object.assign({}, data);
+
+
+		settings = null;
+
+	};
+
+
+	Composite.prototype = {
+
+		// deserialize: function(blob) {},
+
+		serialize: function() {
+
+			return {
+				'constructor': 'lychee.Specification',
+				'arguments': []
+			};
+
+		}
+
+	};
+
+
+	return Composite;
+
+});
+